Use more accurate integer types

+++ src/ec_debug.h
@@ -8,9 +8,9 @@ extern EC_VTable* EC_Debug_Controller;
 
 Error* EC_Debug_Open();
 void   EC_Debug_Close();
-Error* EC_Debug_WriteByte(int, uint8_t);
-Error* EC_Debug_WriteWord(int, uint16_t);
-Error* EC_Debug_ReadByte(int, uint8_t*);
-Error* EC_Debug_ReadWord(int, uint16_t*);
+Error* EC_Debug_WriteByte(uint8_t, uint8_t);
+Error* EC_Debug_WriteWord(uint8_t, uint16_t);
+Error* EC_Debug_ReadByte(uint8_t, uint8_t*);
+Error* EC_Debug_ReadWord(uint8_t, uint16_t*);
 
 #endif
