+++ src/ec_dummy.h
@@ -7,9 +7,9 @@ extern EC_VTable EC_Dummy_VTable;
 
 Error* EC_Dummy_Open();
 void   EC_Dummy_Close();
-Error* EC_Dummy_WriteByte(int, uint8_t);
-Error* EC_Dummy_WriteWord(int, uint16_t);
-Error* EC_Dummy_ReadByte(int, uint8_t*);
-Error* EC_Dummy_ReadWord(int, uint16_t*);
+Error* EC_Dummy_WriteByte(uint8_t, uint8_t);
+Error* EC_Dummy_WriteWord(uint8_t, uint16_t);
+Error* EC_Dummy_ReadByte(uint8_t, uint8_t*);
+Error* EC_Dummy_ReadWord(uint8_t, uint16_t*);
 
 #endif
